feat: state로 isSearchBarFullSize 변경위한 초기 세팅
- 두번 눌러야 적용되는 오류
