+++ frontend/src/components/Header/Header.tsx
@@ -1,77 +1,26 @@
-import { useContext, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 
-import { Grid } from "@mui/material";
 import Box from "@mui/material/Box";
 
 import { SearchBarStateContext } from "contexts/contexts";
-import MENUS from "mockData/menus";
-import Link from "router/Link";
 
-import GNB from "./GNB/GNB";
+import ChildNodes from "./ChildNodes";
 import {
   indexHeaderStyle,
   HeaderContainer,
   miniHeaderStyle,
 } from "./Header.style";
-import SearchBar from "./SearchBar/SearchBar";
-import UserMenu from "./UserMenu/UserMenu";
-
-const LogoArea = () => {
-  return (
-    <Grid container item xs={2} justifyContent="left">
-      <h1 style={{ margin: "auto 0" }}>
-        <Link to="index">LOGO</Link>
-      </h1>
-    </Grid>
-  );
-};
-
-const GNBArea = () => {
-  return (
-    <GNB
-      menuData={MENUS}
-      container
-      rowSpacing={2}
-      item
-      xs={8}
-      justifyContent="center"
-    />
-  );
-};
-
-const headerMiddleItem = {
-  fullSize: <GNBArea />,
-  miniSize: (
-    <Grid item container xs={8} justifyContent="center" alignItems="center">
-      <SearchBar />
-    </Grid>
-  ),
-};
-
-const ChildNodes = () => {
-  const { isSearchBarFullSize } = useContext(SearchBarStateContext)!;
-  console.log(isSearchBarFullSize, "차일드노드");
-  return (
-    <>
-      <Grid
-        container
-        spacing={2}
-        columns={12}
-        sx={{ height: ({ elementSize }) => elementSize.fullSize }}
-      >
-        <LogoArea />
-        {isSearchBarFullSize
-          ? headerMiddleItem.fullSize
-          : headerMiddleItem.miniSize}
-        <UserMenu />
-      </Grid>
-      {isSearchBarFullSize && <SearchBar />}
-    </>
-  );
-};
 
 const Header = () => {
   const [isSearchBarFullSize, setIsSearchBarFullSize] = useState(true);
+
+  useEffect(() => {
+    console.log("headerUseEffect", isSearchBarFullSize);
+  }, [isSearchBarFullSize]);
+
+  // const changeIsSearchBarFullSize = () => {
+  //   setIsSearchBarFullSize((bool) => !bool);
+  // };
 
   return (
     <SearchBarStateContext.Provider
