+++ frontend/src/contexts/contexts.ts
@@ -27,6 +27,7 @@ interface PriceRangeContextType {
 }
 
 interface SearchBarStateContextType {
+  // isSearchBarFullSize: React.MutableRefObject<boolean>;
   isSearchBarFullSize: boolean;
   setIsSearchBarFullSize: React.Dispatch<React.SetStateAction<boolean>>;
 }
