+++ frontend/src/router/Link.tsx
@@ -5,7 +5,7 @@ import { RouterContext } from "./Contexts";
 const { history, location } = window;
 
 const pushHistory = ({ path, state }: PushHistoryProps): void => {
-  // history.pushState(state, title, url)
+  // history.pushState(state, title, url);
   history.pushState(state, path, `${location.origin}${path}`);
 };
 
@@ -17,7 +17,7 @@ const Link = ({
 }: LinkProps): JSX.Element => {
   const { setPage } = useContext(RouterContext);
 
-  const href = to !== "index" ? `/${to}` : `/`;
+  const href = to === "index" ? `/` : `/${to}`;
 
   const handleClick = (e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
     e.preventDefault();
